일러스트 오류 + 경로 수정, 푸터 경로 설정 완료

+++ src/main/java/com/team1/stelling/service/IllustService.java
@@ -48,7 +48,7 @@ public class IllustService {
         return illustRepository.findByIllustHashTagContaining(keyword, pageable);
     }
 
-    public int getLikeTotal(Long userNumber){
+    public Long getLikeTotal(Long userNumber){
         return illustRepository.findByIllustLikeTotal(userNumber);
     }
